flask/Flask_Auth/bacancytechnology/requirements.txt
Flask
pyjwt
datetime
uuid
Flask-SQLAlchemy